Add weekly report period to PeriodicReport (#333)
* Add weekly report period to PeriodicReport

Added a new weekly value to ReportPeriod and updated PeriodicReport to
accommodate it. Updated AT_PeriodicReport to test for weekly reports and
updated other failing tests to now account for the weekly
ReportPeriod value.

* Update weekly logic in PeriodicReport

Updated the weekly logic in PeriodicReport to now have weekly reports
be flushed on an interval of day 7, day 14, day 21 etc regardless of
what the sim start time was. With the old logic, we were just adding 7
days to whatever the sim start time was, which if the sim start time was
3.5 then the flush interval would be day 10, day 17, day 24, etc.